+++ src/main/java/coffeemeet/server/common/annotation/Login.java
@@ -1,4 +1,4 @@
-package coffeemeet.server.common.presentation.annotation;
+package coffeemeet.server.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
