+++ src/main/java/coffeemeet/server/common/presentation/resolver/UserArgumentResolver.java
@@ -5,7 +5,7 @@ import static coffeemeet.server.auth.exception.AuthErrorCode.AUTHENTICATION_FAIL
 import coffeemeet.server.auth.domain.JwtTokenProvider;
 import coffeemeet.server.auth.domain.RefreshToken;
 import coffeemeet.server.auth.implement.RefreshTokenQuery;
-import coffeemeet.server.common.presentation.annotation.Login;
+import coffeemeet.server.common.annotation.Login;
 import coffeemeet.server.common.execption.InvalidInputException;
 import coffeemeet.server.user.presentation.dto.AuthInfo;
 import jakarta.servlet.http.HttpServletRequest;
