+++ src/main/java/coffeemeet/server/user/presentation/UserController.java
@@ -1,7 +1,7 @@
 package coffeemeet.server.user.presentation;
 
 import coffeemeet.server.auth.domain.AuthTokens;
-import coffeemeet.server.common.presentation.annotation.Login;
+import coffeemeet.server.common.annotation.Login;
 import coffeemeet.server.common.util.FileUtils;
 import coffeemeet.server.user.domain.OAuthProvider;
 import coffeemeet.server.user.presentation.dto.AuthInfo;
